--- Teams/05 Microsoft Teams apps - Advanced Techniques/Demos/02 - teams-bot2/teams-bot2.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26430.15
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "teams-bot2", "teams-bot2\teams-bot2.csproj", "{1B17118A-E7D9-4C5D-A667-B5BAF677BE31}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1B17118A-E7D9-4C5D-A667-B5BAF677BE31}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1B17118A-E7D9-4C5D-A667-B5BAF677BE31}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1B17118A-E7D9-4C5D-A667-B5BAF677BE31}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1B17118A-E7D9-4C5D-A667-B5BAF677BE31}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
