--- Teams/05 Microsoft Teams apps - Advanced Techniques/Demos/02 - teams-bot2/teams-bot2/Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="teams_bot2.WebApiApplication" Language="C#" %>
