+++ Teams/05 Microsoft Teams apps - Advanced Techniques/Demos/02-messaging-extension/solution/teams-m5-bot.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28010.2019
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "teams-m5-bot", "teams-m5-bot\teams-m5-bot.csproj", "{89040323-EEB7-476E-87D9-E8FB52F76D33}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{89040323-EEB7-476E-87D9-E8FB52F76D33}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{89040323-EEB7-476E-87D9-E8FB52F76D33}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{89040323-EEB7-476E-87D9-E8FB52F76D33}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{89040323-EEB7-476E-87D9-E8FB52F76D33}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {0DBF6918-2275-4C2F-B4B5-1272F95D7383}
+	EndGlobalSection
+EndGlobal
