+++ Teams/05 Microsoft Teams apps - Advanced Techniques/Demos/02-messaging-extension/solution/teams-m5-bot/Dialogs/RootDialog.cs
@@ -1,0 +1,49 @@
+﻿using System.Threading;
+using System;
+using System.Threading.Tasks;
+using Microsoft.Bot.Builder.Dialogs;
+using Microsoft.Bot.Connector;
+using Microsoft.Bot.Connector.Teams;
+
+namespace teams_m5_bot.Dialogs
+{
+  [Serializable]
+  public class RootDialog : IDialog<object>
+  {
+    public async Task StartAsync(IDialogContext context)
+    {
+      context.Wait(MessageReceivedAsync);
+    }
+
+    public virtual async Task MessageReceivedAsync(IDialogContext context, IAwaitable<IMessageActivity> result)
+    {
+      var activity = await result as Activity;
+
+      if (activity.Text.ToLower().Contains("card"))
+      {
+        await context.Forward(new Dialogs.CardsDialog(), this.ResumeAfterCardsDialog, activity, CancellationToken.None);
+      }
+      else
+      {
+        // calculate something for us to return
+        int length = (activity.Text ?? string.Empty).Length;
+
+        ConnectorClient connector = new ConnectorClient(new Uri(activity.ServiceUrl));
+        Activity reply = activity.CreateReply($"You sent {activity.Text} which was {activity.Text.Length} characters");
+        var msgToUpdate = await connector.Conversations.ReplyToActivityAsync(reply);
+
+        // Simulate external activity
+        System.Threading.Thread.Sleep(5000);
+
+        var withoutMentions = activity.GetTextWithoutMentions();
+
+        Activity updatedReply = activity.CreateReply($"Actually, removing the @ mention, it was {withoutMentions.Length} characters");
+        await connector.Conversations.UpdateActivityAsync(reply.Conversation.Id, msgToUpdate.Id, updatedReply);
+      }
+    }
+    private async Task ResumeAfterCardsDialog(IDialogContext context, IAwaitable<IMessageActivity> result)
+    {
+      context.Wait(this.MessageReceivedAsync);
+    }
+  }
+}
