+++ Teams/05 Microsoft Teams apps - Advanced Techniques/Demos/02-messaging-extension/solution/teams-m5-bot/Global.asax
@@ -1,0 +1,1 @@
+﻿<%@ Application Codebehind="Global.asax.cs" Inherits="teams_m5_bot.WebApiApplication" Language="C#" %>
