+++ Teams/05 Microsoft Teams apps - Advanced Techniques/Demos/02-messaging-extension/solution/teams-m5-bot/Web.Debug.config
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<!-- For more information on using web.config transformation visit http://go.microsoft.com/fwlink/?LinkId=125889 -->
+<!-- For more information on using web.config transformation visit https://go.microsoft.com/fwlink/?LinkId=125889 -->
 
 <configuration xmlns:xdt="http://schemas.microsoft.com/XML-Document-Transform">
   <!--
@@ -15,7 +15,6 @@
     </connectionStrings>
   -->
   <system.web>
-    <compilation xdt:Transform="RemoveAttributes(debug)" />
     <!--
       In the example below, the "Replace" transform will replace the entire 
       <customErrors> section of your web.config file.
