+++ Teams/05 Microsoft Teams apps - Advanced Techniques/Demos/02-messaging-extension/solution/teams-m5-bot/packages.config
@@ -1,0 +1,38 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Autofac" version="4.6.0" targetFramework="net46" />
+  <package id="Chronic.Signed" version="0.3.2" targetFramework="net46" />
+  <package id="EntityFramework" version="6.1.3" targetFramework="net46" />
+  <package id="Microsoft.AspNet.WebApi" version="5.2.3" targetFramework="net46" />
+  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net46" />
+  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net46" />
+  <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.3" targetFramework="net46" />
+  <package id="Microsoft.Azure.DocumentDB" version="1.22.0" targetFramework="net46" />
+  <package id="Microsoft.Azure.KeyVault.Core" version="1.0.0" targetFramework="net46" />
+  <package id="Microsoft.Bot.Builder" version="3.15.2.2" targetFramework="net46" />
+  <package id="Microsoft.Bot.Builder.Azure" version="3.15.2.2" targetFramework="net46" />
+  <package id="Microsoft.Bot.Builder.History" version="3.15.2.2" targetFramework="net46" />
+  <package id="Microsoft.Bot.Connector" version="3.15.2.2" targetFramework="net46" />
+  <package id="Microsoft.Bot.Connector.Teams" version="0.9.0" targetFramework="net46" />
+  <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="1.0.4" targetFramework="net46" />
+  <package id="Microsoft.Data.Edm" version="5.7.0" targetFramework="net46" />
+  <package id="Microsoft.Data.OData" version="5.7.0" targetFramework="net46" />
+  <package id="Microsoft.Data.Services.Client" version="5.7.0" targetFramework="net46" />
+  <package id="Microsoft.IdentityModel.Logging" version="1.1.4" targetFramework="net46" />
+  <package id="Microsoft.IdentityModel.Protocol.Extensions" version="1.0.4.403061554" targetFramework="net46" />
+  <package id="Microsoft.IdentityModel.Protocols" version="2.1.4" targetFramework="net46" />
+  <package id="Microsoft.IdentityModel.Protocols.OpenIdConnect" version="2.1.4" targetFramework="net46" />
+  <package id="Microsoft.IdentityModel.Tokens" version="5.1.4" targetFramework="net46" />
+  <package id="Microsoft.Net.Compilers" version="2.0.1" targetFramework="net46" developmentDependency="true" />
+  <package id="Microsoft.Rest.ClientRuntime" version="2.3.8" targetFramework="net46" />
+  <package id="Microsoft.WindowsAzure.ConfigurationManager" version="3.2.3" targetFramework="net46" />
+  <package id="Newtonsoft.Json" version="10.0.2" targetFramework="net46" />
+  <package id="Polly-Signed" version="5.3.1" targetFramework="net46" />
+  <package id="System.Collections" version="4.3.0" targetFramework="net46" />
+  <package id="System.IdentityModel.Tokens.Jwt" version="5.1.4" targetFramework="net46" />
+  <package id="System.Linq" version="4.3.0" targetFramework="net46" />
+  <package id="System.Spatial" version="5.7.0" targetFramework="net46" />
+  <package id="System.Threading.Tasks" version="4.3.0" targetFramework="net46" />
+  <package id="System.Web.Http.Common" version="4.0.20126.16343" targetFramework="net46" />
+  <package id="WindowsAzure.Storage" version="7.2.1" targetFramework="net46" />
+</packages>
